Add execution=consistent|performant option to the methods that support it
If execution flag is not specified on method that supports it then it will
default to supplying execution=consistent to the Delta API. The execution=consistent
parameter will instruct the Delta API to enforce consistent writes meaning the
response will not be returned until all supported indexes have been updated, or it fails.
Setting the execution parameter to performant will result in the default Nexus Delta
behaviour which is to return immediately after initial success prior to indexes
being updated.