Add support to the AccessControl component to accept multiple paths
The AccessControl component took a single path string along with a
permissions array and would determine if the those permissions
were granted on the path or not. It now supports passing in an
array of paths and it will determine if the array of permissions
exists on each and everyone of the paths or not.